requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
email-validator
python-dotenv
gunicorn
